feeders_rfid/read from sql/requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
collections-extended==1.0.2
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
imageio==2.5.0
imutils==0.5.2
iso8601==0.1.12
isort==4.3.15
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1
matplotlib==3.0.3
mccabe==0.6.1
mkl-fft==1.0.12
mkl-random==1.0.2
mock==2.0.0
more-itertools==7.0.0
networkx==2.2
numpy==1.16.3
opencv-python==4.0.0.21
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
pluggy==0.9.0
protobuf==3.7.1
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
pygame==1.9.6
pylint==2.3.1
pyparsing==2.4.0
PySimpleGUI==3.38.0
pytest==4.3.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.2
PyYAML==5.1
random2==1.0.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
serial==0.0.97
six==1.12.0
sounddevice==0.3.13
SoundFile==0.10.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
Werkzeug==0.15.2
wrapt==1.11.1
